5.15

+++ Frito1/RealmObjects.swift
@@ -11,7 +11,7 @@ import RealmSwift
 class Shouhin: Object {
     @objc dynamic var name = ""
     @objc dynamic var category = ""
-    @objc dynamic var number: Float = 0.0
+    @objc dynamic var number: Double = 0.0
     @objc dynamic var id: Int = 1
     
     override static func primaryKey() -> String {
